fix dynamic reload and positions, cleanup
