+++ src/lib/domains/leaderboard/components/Competition/DappCompetition/Chillblazer/ChillblazerHeader.svelte
@@ -10,6 +10,8 @@
   import type { PaginationInfo } from '$shared/dto/CommonPageApiResponse';
   import { classNames } from '$shared/utils/classNames';
 
+  import { chillblazerDetails } from '../Chillblazer/chillblazerDetails';
+
   const loadLeaderboardData = getContext<LoadLeaderboardDataType>('loadCompetitionLeaderboardData');
   const pageInfo = getContext<PaginationInfo<DappLeaderboardItem>>('dappsCompetitionPageInfo');
 
@@ -18,6 +20,8 @@
   };
 
   export let lastUpdated: Date;
+
+  const edition = getContext<number>('chillblazerEdition');
 
   const wrapperCLasses = classNames('f-col', 'relative', 'mt-[40px]', 'w-full', 'items-center');
 
@@ -90,30 +94,22 @@
     'xl:my-[40px]',
   );
 
-  const competitionInfo: CompetitionInfo = {
+  // Initialize reactive variables
+  let selectedCompetitionInfo: CompetitionInfo = {
     title: '',
-    description: $t('leaderboard.gaming.banner.description'),
-    prizeTitle: $t('leaderboard.gaming.prize.1'),
-    prizeSubtitle: $t('leaderboard.chillblazers.prize.2'),
-    prizes: [
-      {
-        image: '/first.svg',
-        amount: $t('leaderboard.chillblazers.prize_breakdown.first.amount'),
-      },
-      {
-        image: '/second.svg',
-        amount: $t('leaderboard.chillblazers.prize_breakdown.second.amount'),
-      },
-      {
-        image: '/third.svg',
-        amount: $t('leaderboard.chillblazers.prize_breakdown.third.amount'),
-      },
-      {
-        image: '/default-prize.svg',
-        amount: $t('leaderboard.chillblazers.prize_breakdown.fourth.amount'),
-      },
-    ],
+    description: '',
+    prizeTitle: '',
+    prizeSubtitle: '',
+    prizes: [],
+    qualifyingPositions: 3,
   };
+
+  $: description = '';
+
+  $: {
+    selectedCompetitionInfo = chillblazerDetails[edition];
+    description = selectedCompetitionInfo.description;
+  }
 
   const h2Classes = classNames(
     'text-center',
@@ -241,8 +237,7 @@
   <img src="/blazers.svg" alt="Thrillblazers" class={blazersClasses} />
 
   <span class={descriptionClasses}>
-    Chill-Blazer launches Nov 18th. Calling all gaming and SocialFi dApps: Blaze new trails, fuel Taiko's momentum, and
-    seize your chance for legendary rewards!
+    {description}
   </span>
 
   <img src="/header/thrillblazer/runes-right.svg" alt="runes" class={runesRightClasses} />
@@ -252,10 +247,10 @@
     Prize pool
     <span class="text-white align-baseline">&#x25C6</span>
   </h2>
-  <h3 class={h3Classes}>{competitionInfo.prizeSubtitle}</h3>
+  <h3 class={h3Classes}>{selectedCompetitionInfo.prizeSubtitle}</h3>
 
   <div class={priceWrapperClasses}>
-    {#each competitionInfo.prizes as prize}
+    {#each selectedCompetitionInfo.prizes as prize}
       <div class={prizeItemClasses}>
         <img src={prize.image} alt="{prize.amount} TAIKO Tokens" class={prizeImageClasses} />
         <div class={prizeInfoClasses}>
