+++ src/lib/domains/leaderboard/components/Competition/DappCompetition/Chillblazer/ChillblazerLeaderboard.svelte
@@ -1,73 +1,77 @@
-<!-- <script lang="ts">
-  import { setContext } from 'svelte';
+<script lang="ts">
+  import { onMount, setContext } from 'svelte';
   import { t } from 'svelte-i18n';
 
-  import { leaderboardConfig } from '$config';
+  import { browser } from '$app/environment';
   import { CampaignEndedInfoBox } from '$lib/domains/leaderboard/components/CampaignEndedInfoBox';
   import { AbstractLeaderboard, PointScore } from '$lib/domains/leaderboard/components/Template';
   import type { DappLeaderboardItem } from '$lib/domains/leaderboard/dto/dapps.dto';
-  import { chillblazerService } from '$lib/domains/leaderboard/services/LeaderboardServiceInstances';
-  import type { DappLeaderboardPage } from '$lib/domains/leaderboard/types/dapps/types';
+  import { fetchLeaderboard, leaderboardStore } from '$lib/domains/leaderboard/stores/dappCompetitionStore';
+  import { CompetitionType } from '$lib/domains/leaderboard/types/competition/types';
   import type { PaginationInfo } from '$lib/shared/dto/CommonPageApiResponse';
+  import { activeSeason } from '$shared/stores/activeSeason';
   import { getLogger } from '$shared/utils/logger';
 
   import ChillblazerHeader from './ChillblazerHeader.svelte';
 
-  const log = getLogger('DappsLeaderboard');
+  const log = getLogger('ChillblazerLeaderboard');
+
+  // Props from server load
   export let loading = false;
   export let pageInfo: PaginationInfo<DappLeaderboardItem>;
-  export let season: number;
+  export let edition: number;
 
-  const endedSeasons: number[] = [2];
+  const currentEdition: number = 4;
 
   $: totalItems = pageInfo?.total || 0;
-  $: pageSize = pageInfo?.size || leaderboardConfig.pageSize;
-
-  $: hasEnded = endedSeasons.includes(season);
+  // For Chillblazer, you might consider the competition ended if the edition is not the current one.
+  $: hasEnded = edition !== currentEdition;
+  $: isLoading = loading;
 
   function handlePageChange(page: number) {
     log('handlePageChange', page);
-    loadLeaderboardData(page);
+    if (browser) fetchLeaderboard(page, CompetitionType.CHILLBLAZER, edition);
   }
 
-  const currentDappCompetitionLeaderboard = chillblazerService.getStore();
-
-  async function loadLeaderboardData(page: number, name = '') {
-    log('loadLeaderboardData', page, name);
+  async function loadLeaderboardData(page: number) {
+    log('loadLeaderboardData', page);
     loading = true;
-    // Fetch the leaderboard data for the given page
-    const args: PaginationInfo<DappLeaderboardItem> = {
-      page,
-      size: pageSize,
-      slug: name,
-      total: totalItems,
-    };
-
-    const leaderboardPage: DappLeaderboardPage = await chillblazerService.fetchCompetitionData(args, season);
-
-    // date from timestamp
-    totalItems = leaderboardPage?.pagination.total || $currentDappCompetitionLeaderboard.items.length;
+    if (browser) await fetchLeaderboard(page, CompetitionType.CHILLBLAZER, edition);
     loading = false;
   }
 
+  let leaderboard;
+  $: leaderboard = $leaderboardStore;
+
+  onMount(() => {
+    if (browser && $activeSeason && pageInfo) {
+      loadLeaderboardData(pageInfo.page);
+    }
+  });
+
   setContext('loadCompetitionLeaderboardData', loadLeaderboardData);
   setContext('dappsCompetitionPageInfo', pageInfo);
+  setContext('chillblazerEdition', edition);
 </script>
 
-<AbstractLeaderboard
-  headers={['No.', 'Dapp', '', 'Points']}
-  data={$currentDappCompetitionLeaderboard.items}
-  showTrophy={true}
-  lastUpdated={new Date($currentDappCompetitionLeaderboard.lastUpdated)}
-  isLoading={loading}
-  {handlePageChange}
-  {totalItems}
-  ended={hasEnded}
-  qualifyingPositions={4}
-  endedComponent={CampaignEndedInfoBox}
-  endTitleText={$t('leaderboard.chillblazers.ended.title')}
-  endDescriptionText={$t('leaderboard.chillblazers.ended.description')}
-  showPagination={true}
-  {season}
-  headerComponent={ChillblazerHeader}
-  scoreComponent={PointScore} /> -->
+{#if edition <= currentEdition}
+  <AbstractLeaderboard
+    headers={['No.', 'Dapp', '', 'Points']}
+    season={$activeSeason}
+    data={leaderboard.items}
+    showTrophy={true}
+    lastUpdated={new Date(leaderboard.lastUpdated)}
+    {isLoading}
+    {handlePageChange}
+    {totalItems}
+    ended={hasEnded}
+    qualifyingPositions={4}
+    endedComponent={CampaignEndedInfoBox}
+    endTitleText={$t(`leaderboard.chillblazers.${edition}.ended.title`)}
+    endDescriptionText={$t(`leaderboard.chillblazers.${edition}.ended.description`)}
+    showPagination={true}
+    headerComponent={ChillblazerHeader}
+    scoreComponent={PointScore} />
+{:else}
+  No data
+{/if}
