+++ src/lib/domains/leaderboard/components/Competition/DappCompetition/Thrillblazer/Header/Header.svelte
