--- src/lib/domains/leaderboard/components/Competition/DefiCompetition/DefiEndedComponent.svelte
@@ -1,8 +1,0 @@
-<script lang="ts">
-  import { CampaignEndedInfoBox } from '$lib/domains/leaderboard/components/CampaignEndedInfoBox';
-
-  export let title: string;
-  export let description: string;
-</script>
-
-<CampaignEndedInfoBox {title} {description} />
