--- src/lib/domains/leaderboard/components/Competition/DefiCompetition/index.ts
@@ -1,2 +1,0 @@
-export { default as DefiCompetitionInformation } from './DefiCompetitionInformation.svelte';
-export { default as DefiEndedComponent } from './DefiEndedComponent.svelte';
