--- src/lib/domains/leaderboard/components/Competition/GamingCompetition/index.ts
@@ -1,1 +1,0 @@
-export { default as GamingCompetitionInformation } from './GamingCompetitionInformation.svelte';
