+++ src/lib/domains/leaderboard/components/index.ts
@@ -1,4 +1,2 @@
 export { default as BridgeLeaderboard } from './BridgeLeaderboard.svelte';
-export { default as DeFiDappLeaderboard } from './DeFiDappLeaderboard.svelte';
-export { default as GamingLeaderboard } from './GamingLeaderboard.svelte';
 export { default as LastUpdated } from './LastUpdated.svelte';
