--- src/routes/leaderboard/s1/defi/+page.svelte
@@ -1,21 +1,0 @@
-<script lang="ts">
-  import { t } from 'svelte-i18n';
-
-  import { goto } from '$app/navigation';
-  import DeFiDappLeaderboard from '$lib/domains/leaderboard/components/DeFiDappLeaderboard.svelte';
-  import { ActionButton } from '$shared/components/Button';
-  import { Page } from '$shared/components/Page';
-</script>
-
-<svelte:head>
-  <title>Taiko Trailblazer - DeFi Leaderboard</title>
-</svelte:head>
-
-<Page>
-  <DeFiDappLeaderboard />
-  <div class="w-full flex justify-center mt-[58px]">
-    <ActionButton class="max-w-[280px]" priority="primary" on:click={() => goto('/leaderboard/dapp')} withArrow>
-      {$t('buttons.leaderboard.dapp')}
-    </ActionButton>
-  </div>
-</Page>
