--- src/routes/leaderboard/s2/gaming/+page.ts
@@ -1,33 +1,0 @@
-import { browser } from '$app/environment';
-import { leaderboardConfig } from '$config';
-import type { GamingLeaderboardItem } from '$lib/domains/leaderboard/dto/gaming.dto';
-import { gamingLeaderboardService } from '$lib/domains/leaderboard/services/LeaderboardServiceInstances';
-import type { PaginationInfo } from '$lib/shared/dto/CommonPageApiResponse';
-
-export const load = async () => {
-  let loading = true;
-  let pageInfo: PaginationInfo<GamingLeaderboardItem> = {
-    page: 0,
-    size: leaderboardConfig.pageSize,
-    first: 0,
-    last: 1,
-    total: 0,
-  };
-
-  if (browser) {
-    try {
-      const page = await gamingLeaderboardService.getGamingLeaderboardData(pageInfo, 2);
-      if (page) {
-        pageInfo = page.pagination;
-      }
-    } catch (error) {
-      console.error('Error loading leaderboard data:', error);
-    } finally {
-      loading = false;
-    }
-  }
-  return {
-    pageInfo,
-    loading,
-  };
-};
